feat(shipping): CHECKOUT-8596 address review comments
